backend/requirements.txt
fastapi
uvicorn[standard]
litellm
python-multipart
httpx
python-dotenv
pydantic